FileAPI/file/worker-read-file-constructor.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Read file info in Worker</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/FileAPI/#file">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

async_test(function() {
  var file = new File(["bits"], "dummy", { 'type': 'text/plain', lastModified: 42 });

  var worker = new Worker("../support/worker-read-file-constructor.js");

  worker.onmessage = this.step_func(function(event) {
    assert_equals(event.data.name, file.name, "file name");
    assert_equals(event.data.content, "bits", "file content");
    assert_equals(event.data.lastModified, file.lastModified, "file lastModified");
    this.done();
  });

  worker.onerror = this.step_func(function(event) {
    assert_unreached(event.message);
  });
  worker.postMessage(file);

}, "FileReader in Worker")

</script>
